+++ packages/core/src/components/RichDescription.tsx
@@ -1,7 +1,8 @@
 import type { ReactElement } from 'react';
 import type { FormContextType, Registry, RJSFSchema, StrictRJSFSchema, UiSchema } from '@rjsf/utils';
 import { getTestIds, getUiOptions } from '@rjsf/utils';
-import { Markdown } from 'markdown-to-jsx/react';
+
+import Markdown from './LazyMarkdown';
 
 const TEST_IDS = getTestIds();
 
@@ -31,11 +32,7 @@ export default function RichDescription<
   const uiOptions = getUiOptions<T, S, F>(uiSchema, globalUiOptions);
 
   if (uiOptions.enableMarkdownInDescription && typeof description === 'string') {
-    return (
-      <Markdown options={{ disableParsingRawHTML: true }} data-testid={TEST_IDS.markdown}>
-        {description}
-      </Markdown>
-    );
+    return <Markdown data-testid={TEST_IDS.markdown}>{description}</Markdown>;
   }
   return description;
 }