+++ packages/core/src/components/templates/UnsupportedField.tsx
@@ -1,6 +1,7 @@
 import type { FormContextType, RJSFSchema, StrictRJSFSchema, UnsupportedFieldProps } from '@rjsf/utils';
 import { TranslatableString } from '@rjsf/utils';
-import { Markdown } from 'markdown-to-jsx/react';
+
+import Markdown from '../LazyMarkdown';
 
 /** The `UnsupportedField` component is used to render a field in the schema is one that is not supported by
  * react-jsonschema-form.
@@ -28,7 +29,7 @@ function UnsupportedField<T = any, S extends StrictRJSFSchema = RJSFSchema, F ex
   return (
     <div className='unsupported-field'>
       <p>
-        <Markdown options={{ disableParsingRawHTML: true }}>{translateString(translateEnum, translateParams)}</Markdown>
+        <Markdown>{translateString(translateEnum, translateParams)}</Markdown>
       </p>
       {schema && <pre>{JSON.stringify(schema, null, 2)}</pre>}
     </div>