+++ packages/core/src/components/widgets/FileWidget.tsx
@@ -9,7 +9,8 @@ import type {
   WidgetProps,
 } from '@rjsf/utils';
 import { getTemplate, TranslatableString, useFileWidgetProps } from '@rjsf/utils';
-import { Markdown } from 'markdown-to-jsx/react';
+
+import Markdown from '../LazyMarkdown';
 
 function FileInfoPreview<T = any, S extends StrictRJSFSchema = RJSFSchema, F extends FormContextType = any>({
   fileInfo,