+++ packages/core/src/index.ts
@@ -1,5 +1,7 @@
 import type { FormProps, FormState, IChangeEvent } from './components/Form';
 import Form from './components/Form';
+import type { LazyMarkdownProps } from './components/LazyMarkdown';
+import LazyMarkdown, { preloadMarkdown } from './components/LazyMarkdown';
 import type { RichDescriptionProps } from './components/RichDescription';
 import RichDescription from './components/RichDescription';
 import type { RichHelpProps } from './components/RichHelp';
@@ -15,11 +17,21 @@ export type {
   FormProps,
   FormState,
   IChangeEvent,
+  LazyMarkdownProps,
   ThemeProps,
   RichDescriptionProps,
   RichHelpProps,
   SchemaExamplesProps,
 };
 
-export { withTheme, getDefaultRegistry, getTestRegistry, RichDescription, RichHelp, SchemaExamples };
+export {
+  withTheme,
+  getDefaultRegistry,
+  getTestRegistry,
+  LazyMarkdown,
+  preloadMarkdown,
+  RichDescription,
+  RichHelp,
+  SchemaExamples,
+};
 export default Form;